# Improve perf when building Ruby arrays from Rust