Applies Mat's suggestion from #169, essentially using RArray's FromIterator implementation which appends to the RArray in batch instead of `push`ing in a loop.

Before:
```
Call host func (4 args):   296182.4 i/s
Call host func (16 args):   239771.2 i/s - 1.24x  slower
Call host func (64 args):   105414.9 i/s - 2.81x  slower
Call host func (128 args):    44266.8 i/s - 6.69x  slower
Call host func (256 args):    13739.3 i/s - 21.56x  slower
```

After:
```
Call host func (4 args):   302802.0 i/s
Call host func (16 args):   250214.7 i/s - 1.21x  slower
Call host func (64 args):   114372.2 i/s - 2.65x  slower
Call host func (128 args):    47095.3 i/s - 6.43x  slower
Call host func (256 args):    14555.6 i/s - 20.80x  slower
```

Closes #169